[FEATURE] Add a way for TYPO3CR contents to be migrated
Adds services to migrate specific content in your TYPO3CR to
another state, like changing node types or properties globally.

The change comes with a first migration:

Text and TextWithImage elements are migrated, TwoColumn elements
are not, because the needed transformation is not yet supported by
the node repository (Rename node path).

Change-Id: Iffba5fb6078ad971c87fc8a001953db5d4a3865a
Resolves: #40373