[BUGFIX] Fix setRemoved() for proxy nodes
The method Node::setRemoved() has not been implemented for ProxyNode
and thus nodes which are proxy nodes, are not marked as removed when
using that method.

This patch also introduces tests for that part and a new LoggingAspect
which - for now - only logs removal of nodes.

This bug has been introduced in 773160cbd25847049a8b57a35a3e3cfb96651be1

Change-Id: I903e144d037de9698579ee68c634a517f0f0eaa0
Related: #40754