[!!!][FEATURE] Rename ContentType to NodeType
As decided on the code sprint, the type of a node should consistently
be called "Node Type" instead of "Content Type".

This change renames all occurrences of "node type" to "content type".

The frontend and backend works for me as expected; though this should
also be validated by somebody else.

This change is breaking because a database migration needs to be
executed. The DB migration for Postgresql is not yet generated.